* commit 'd6d0b8116abbae1a3d2e096713b5e25bc713d43f':
  Fix combination bug;
  Fix combiantion bug;
  Modify trendfragment icon;
  Modify mainpage combiantion item click;
  Add mainpage my combination gridview;

+++ src/com/dkhs/portfolio/ui/adapter/AdjustHistoryAdapter.java
@@ -20,6 +20,7 @@ import android.widget.TextView;
 
 import com.dkhs.portfolio.R;
 import com.dkhs.portfolio.bean.PositionDetail.PositionAdjustBean;
+import com.dkhs.portfolio.utils.StringFromatUtils;
 import com.dkhs.portfolio.utils.TimeUtils;
 
 /**
@@ -78,7 +79,7 @@ public class AdjustHistoryAdapter extends BaseAdapter {
                     sb.append(percent);
                     sb.append("%");
                     sb.append("调至");
-                    sb.append(position.getToPercent() * 100 + "%");
+                    sb.append((int) (position.getToPercent() * 100) + "%");
                     bean.content = sb.toString();
                     this.mDataList.add(bean);
                 } else {
@@ -90,7 +91,7 @@ public class AdjustHistoryAdapter extends BaseAdapter {
                     sb.append(percent);
                     sb.append("%");
                     sb.append("调至");
-                    sb.append(position.getToPercent() * 100 + "%");
+                    sb.append((int) (position.getToPercent() * 100) + "%");
                     getLastBean().content = sb.toString();
                 }
             }
